Refactor agenda editing
The intention of these changes was to be able to remove the event listeners that get added when
wanting to edit the details of an agenda. It resulted in needing to refactor the entire
feature due to the nature of removing event listeners that called functions with parameters.

I also wanted to keep the event listeners for saving and cancelling the edit as part of the
parent event listener for the sidebar agenda "tabs". This was to reduce the amount of event
listeners at one time and to prevent having to continuously add and remove these event listeners
on every edit.